Extract functionality to base timeline.
